+++ src/main/java/com/kruemblegard/client/render/CephalariZombieRenderer.java
@@ -2,6 +2,7 @@ package com.kruemblegard.client.render;
 
 import com.kruemblegard.client.render.layer.CephalariZombieDrownedOuterLayer;
 import com.kruemblegard.client.render.layer.CephalariZombieProfessionLayer;
+import com.kruemblegard.client.render.layer.CephalariZombieBodyOverlayLayer;
 import com.kruemblegard.client.render.model.CephalariZombieModel;
 import com.kruemblegard.entity.CephalariZombieEntity;
 
@@ -14,6 +15,7 @@ public class CephalariZombieRenderer extends GeoEntityRenderer<CephalariZombieEn
         super(renderManager, new CephalariZombieModel());
         this.shadowRadius = 0.5F;
 
+        addRenderLayer(new CephalariZombieBodyOverlayLayer(this));
         addRenderLayer(new CephalariZombieDrownedOuterLayer(this));
         addRenderLayer(new CephalariZombieProfessionLayer(this));
     }